Module Brush.Pen

Pens: change the way lines look like in Mlpost

type t = Pen.t

val transform : Transform.t -> t -> t

Apply a transformation to pens

val default : t

Apply a transformation to pens

The default pen; it corresponds to Pen.scale (Num.bp 0.5) Pen.circle

val circle : t

The default pen; it corresponds to Pen.scale (Num.bp 0.5) Pen.circle

A circular pen of diameter 1 bp

val square : t

A circular pen of diameter 1 bp

A pen in form of a square, of length 1 bp

val from_path : Path.t -> t

A pen in form of a square, of length 1 bp

Construct a pen from a closed path

val scale : Num.t -> t -> t

val rotate : float -> t -> t

val shift : Point.t -> t -> t

val yscale : Num.t -> t -> t

val xscale : Num.t -> t -> t

Shortcuts for transformations of pens