Module Mlpost.Concrete

Compute concrete values of numerics, points and paths; not always available

val supported : bool

module CPoint : sig ... end

The module of concrete points

module CPath : sig ... end

Concrete Paths

module CTransform : sig ... end

Compute the concrete value

val float_of_num : Num.t -> float

val cpoint_of_point : Point.t -> CPoint.t

val cpath_of_path : Path.t -> CPath.t

val ctransform_of_transform : Transform.t -> CTransform.t

Compute the baselines of a tex

val baselines : string -> float list

Simple functions for the opposite operations

val num_of_float : float -> Num.t

Same as Num.bp.

val point_of_cpoint : CPoint.t -> Point.t

val path_of_cpath : CPath.t -> Path.t

val transform_of_ctransform : CTransform.t -> Transform.t

Some options (the mlpost tool takes care of them)

val set_verbosity : bool -> unit

(deprecated)

val set_prelude : string -> unit

(deprecated) set_prelude filename uses the prelude of the file filename for compilation of the tex snippets

val set_prelude2 : string option -> unit

(deprecated) set_prelude2 prelude uses this prelude for compilation of the tex snippets

val set_t1disasm : string option -> unit

(deprecated)