Module Mlpost.Helpers

A few helper functions

val dotlabels : ?⁠pos:Command.position -> string list -> Point.t list -> Command.t

val draw_simple_arrow : ?⁠color:Color.t -> ?⁠pen:Pen.t -> ?⁠dashed:Dash.t -> ?⁠style:Path.joint -> ?⁠outd:Path.direction -> ?⁠ind:Path.direction -> ?⁠sep:Num.t -> Point.t -> Point.t -> Command.t

val draw_label_arrow : ?⁠color:Color.t -> ?⁠pen:Pen.t -> ?⁠dashed:Dash.t -> ?⁠style:Path.joint -> ?⁠outd:Path.direction -> ?⁠ind:Path.direction -> ?⁠pos:Command.position -> ?⁠sep:Num.t -> Picture.t -> Point.t -> Point.t -> Command.t

val draw_labelbox_arrow : ?⁠color:Color.t -> ?⁠pen:Pen.t -> ?⁠dashed:Dash.t -> ?⁠style:Path.joint -> ?⁠outd:Path.direction -> ?⁠ind:Path.direction -> ?⁠pos:Command.position -> ?⁠sep:Num.t -> Box.t -> Point.t -> Point.t -> Command.t

val box_arrow : ?⁠within:Box.t -> ?⁠color:Color.t -> ?⁠pen:Pen.t -> ?⁠dashed:Dash.t -> ?⁠style:Path.joint -> ?⁠outd:Path.direction -> ?⁠ind:Path.direction -> ?⁠sep:Num.t -> Box.t -> Box.t -> Command.t

Draw an arrow between two boxes. The options pen, dashed, color change the drawing of the arrow. outd and ind specify the outgoing and ingoing direction. sep specifies the distance of the arrow ends to both boxes. If within is set, the boxes will be searched within the box within.

val box_line : ?⁠within:Box.t -> ?⁠color:Color.t -> ?⁠pen:Pen.t -> ?⁠dashed:Dash.t -> ?⁠style:Path.joint -> ?⁠outd:Path.direction -> ?⁠ind:Path.direction -> ?⁠sep:Num.t -> Box.t -> Box.t -> Command.t

Draw an arrow between two boxes. The options pen, dashed, color change the drawing of the arrow. outd and ind specify the outgoing and ingoing direction. sep specifies the distance of the arrow ends to both boxes. If within is set, the boxes will be searched within the box within.

val box_point_line : ?⁠within:Box.t -> ?⁠color:Color.t -> ?⁠pen:Pen.t -> ?⁠dashed:Dash.t -> ?⁠style:Path.joint -> ?⁠outd:Path.direction -> ?⁠ind:Path.direction -> ?⁠sep:Num.t -> Box.t -> Point.t -> Command.t

val point_box_line : ?⁠within:Box.t -> ?⁠color:Color.t -> ?⁠pen:Pen.t -> ?⁠dashed:Dash.t -> ?⁠style:Path.joint -> ?⁠outd:Path.direction -> ?⁠ind:Path.direction -> ?⁠sep:Num.t -> Point.t -> Box.t -> Command.t

val box_label_arrow : ?⁠within:Box.t -> ?⁠color:Color.t -> ?⁠pen:Pen.t -> ?⁠dashed:Dash.t -> ?⁠style:Path.joint -> ?⁠outd:Path.direction -> ?⁠ind:Path.direction -> ?⁠sep:Num.t -> ?⁠pos:Command.position -> Picture.t -> Box.t -> Box.t -> Command.t

val box_label_line : ?⁠within:Box.t -> ?⁠color:Color.t -> ?⁠pen:Pen.t -> ?⁠dashed:Dash.t -> ?⁠style:Path.joint -> ?⁠outd:Path.direction -> ?⁠ind:Path.direction -> ?⁠sep:Num.t -> ?⁠pos:Command.position -> Picture.t -> Box.t -> Box.t -> Command.t

val box_labelbox_arrow : ?⁠within:Box.t -> ?⁠color:Color.t -> ?⁠pen:Pen.t -> ?⁠dashed:Dash.t -> ?⁠style:Path.joint -> ?⁠outd:Path.direction -> ?⁠ind:Path.direction -> ?⁠sep:Num.t -> ?⁠pos:Command.position -> Box.t -> Box.t -> Box.t -> Command.t

val box_pointer_arrow : ?⁠within:Box.t -> ?⁠color:Color.t -> ?⁠pen:Pen.t -> ?⁠dashed:Dash.t -> ?⁠style:Path.joint -> ?⁠outd:Path.direction -> ?⁠ind:Path.direction -> Box.t -> Box.t -> Command.t

val box_loop : ?⁠within:Box.t -> ?⁠color:Color.t -> ?⁠pen:Pen.t -> ?⁠dashed:Dash.t -> ?⁠style:Path.joint -> ?⁠outd:Path.direction -> ?⁠ind:Path.direction -> ?⁠sep:Num.t -> ?⁠pos:Box.position -> ?⁠dist:float -> ?⁠angle:float -> Box.t -> Command.t

Position pos defaults to `South. angle is the opening angle of the loop, in degrees; it defaults to 90. dist is the relative distance between the box center and the furthest point of the loop (when compared to the distance between the center and the corner pos); it defaults to 2.

val box_label_loop : ?⁠within:Box.t -> ?⁠color:Color.t -> ?⁠pen:Pen.t -> ?⁠dashed:Dash.t -> ?⁠style:Path.joint -> ?⁠outd:Path.direction -> ?⁠ind:Path.direction -> ?⁠sep:Num.t -> ?⁠pos:Box.position -> ?⁠dist:float -> ?⁠angle:float -> Picture.t -> Box.t -> Command.t