Module Mlpost.Plot

A simple and limited way of plotting functions from int to int.

type skeleton

The abstract skeleton for grids, axes and functions

val mk_skeleton : int -> int -> Num.t -> Num.t -> skeleton

mk_skeleton w h dx dy builds a skeleton of width w and height h, each cell being dx units wide and dy units high.

type labels = int -> Num.t -> Picture.t option

type ticks = (Num.t * Pen.t) option

type drawing =

| Stepwise

| Normal

val draw_grid : ?⁠hdash:(int -> Dash.t) -> ?⁠vdash:(int -> Dash.t) -> ?⁠hpen:(int -> Pen.t) -> ?⁠vpen:(int -> Pen.t) -> ?⁠color:Color.t -> skeleton -> Command.t

val draw_axes : ?⁠hpen:Pen.t -> ?⁠vpen:Pen.t -> ?⁠hlabel:labels -> ?⁠vlabel:labels -> ?⁠ticks:ticks -> ?⁠closed:bool -> ?⁠hcaption:Picture.t -> ?⁠vcaption:Picture.t -> skeleton -> Command.t

val draw_simple_axes : ?⁠hpen:Pen.t -> ?⁠vpen:Pen.t -> string -> string -> skeleton -> Command.t

val draw_func : ?⁠pen:Pen.t -> ?⁠drawing:drawing -> ?⁠style:Path.joint -> ?⁠dashed:Dash.t -> ?⁠color:Color.t -> ?⁠label:(Picture.t * Command.position * int) -> ?⁠from_x:int -> ?⁠to_x:int -> (int -> float) -> skeleton -> Command.t