Module Mlpost.Radar

Radar diagrams.

val stack : ?⁠radius:Num.t -> ?⁠color:Color.t list -> ?⁠pen:Pen.t -> ?⁠style:Dash.t list -> ?⁠ticks:float -> ?⁠label:string list -> ?⁠scale:float list -> float list list -> Picture.t

stack l builds a picture from a list l of floating-point lists. The radars are all drawn on the same picture. Each sublist represents one radar datas. All sublists must have the same length.

parameter radius

The radius of the whole picture

parameter pen

The pen used to draw the radars

parameter color

The colors used to draw each radar; it is used circularly

parameter style

The dash-styles used to draw each radar; it is used circularly

parameter ticks

The interval between each ticks along the axes (relative to the values)

parameter scale

The size of every axe, relative to the values; when not specified, the maximal value along each axe is chosen

val compare : ?⁠radius:Num.t -> ?⁠color:Color.t list -> ?⁠fill:bool -> ?⁠pen:Pen.t -> ?⁠style:Dash.t list -> ?⁠ticks:float -> ?⁠label:string list -> ?⁠scale:float list -> float list list -> Picture.t list

stack l builds a list of pictures from a list l of floating-point lists. Each picture represents one radar, and all picture have the same size. Each sublist represents one radar datas, and all sublists must have the same length. For most optional arguments, see function stack above.

parameter fill

Fill the radar with its color.