Module Tree_adv.Place

Parameters

X : Signature.Boxlike

Signature

val gen_place : place:(Box.t t -> Box.t) -> X.t t -> X.t t

This is a generic function for placing trees, provided that the user can give us the following functions:

parameter place

a function which knows how to place a tree of boxes - it should return a box where all the boxes of the input tree appear.

val place : ?⁠ls:Num.t -> ?⁠cs:Num.t -> ?⁠valign:Box.position -> ?⁠halign:Box.position -> X.t t -> X.t t

This is an instance of gen_place using the tree drawing algorithm from the module Tree.