Module Mlpost__.Fonts

type type1 = private {

glyphs_tag : int;

glyphs_ft : Mlpost__.Mlpost_ft.t;

glyphs_enc : int -> int;

slant : float option;

extend : float option;

glyphs_ratio_cm : float;

}

type vf = private {

vf_design_size : float;

vf_font_map : Mlpost__.Dvi_util.font_def Mlpost__.Dvi_util.Int32Map.t;

vf_chars : Mlpost__.Dvi.command list Mlpost__.Dvi_util.Int32H.t;

}

type glyphs =

| Type1 of type1

| VirtualFont of vf

type t

the type of a loaded font

val load_font : Mlpost__.Dvi_util.font_def -> float -> t

load_font def f loads font def scaled by f

val metric : t -> Mlpost__.Tfm.t

Obtain the font metric

val tex_name : t -> string

get the name of the font as used by TeX

val ratio_cm : t -> float

The font ratio, in cm

val glyphs : t -> glyphs

val char_width : t -> int -> float

val char_height : t -> int -> float

val char_depth : t -> int -> float

get information about the ith char of the font

val char_dims : t -> int -> float * float * float

width, height, depth of the ith char

val scale : t -> float -> float

scale t f scale the given float f by ratio_cm t

val design_size : t -> float

the design size of the font