Module Mlpost__File

module LowLevel : sig ... end

val dir_sep : char

val dir_sep_string : string

module Dir : sig ... end

type t

a file name, including directory information

val mk : ?⁠dir:Dir.t -> string -> string -> t

give a directory, a base name and an extension, and obtain a file name

val from_string : string -> t

val to_string : t -> string

val debug_to_string : t -> string

val place : Dir.t -> t -> t

replace the current directory information of the file by the one given

val concat : Dir.t -> t -> t

concat directory information given to the one of the file

val append : t -> string -> t

val prepend : t -> string -> t

append string to file name - do not use this for file extensions

val move : t -> t -> unit

move a file to another place

val copy : t -> t -> unit

val read_from : t -> (Stdlib.in_channel -> 'a) -> 'a

val compare : t -> t -> int

val basename : t -> string

val extension : t -> string

val dir : t -> Dir.t

val exists : t -> bool

val rm : t -> unit

val open_out : t -> Stdlib.out_channel

val open_in : t -> Stdlib.in_channel

val open_in_gen : Stdlib.open_flag list -> int -> t -> Stdlib.in_channel

val clear_dir : t -> t

val set_ext : t -> string -> t

clear extension if passed empty string

val write_to : t -> (Stdlib.out_channel -> 'a) -> 'a

val write_to_formatted : t -> (Stdlib.Format.formatter -> 'a) -> 'a

module Map : Stdlib.Map.S with type Map.key = t