Module Mlpost__Metapath_lib.Approx

val lineto : point list -> Mlpost__.Spline_lib.path

val fullcircle : float -> Mlpost__.Spline_lib.path

fullcircle l is the circle of diameter l centered on (0, 0)

val halfcirle : float -> Mlpost__.Spline_lib.path

fullcircle l is the circle of diameter l centered on (0, 0)

halfcircle l is the upper half of a fullcircle of diameter l

val quartercircle : float -> Mlpost__.Spline_lib.path

halfcircle l is the upper half of a fullcircle of diameter l

quartercircle l is the first quadrant of a circle of diameter l

val unitsquare : float -> Mlpost__.Spline_lib.path

quartercircle l is the first quadrant of a circle of diameter l

unitsquare l : the path (0,0)--(l,0)--(l,l)--(0,l)--cycle