Module Mlpost__Metapost

val generate : ?⁠prelude:string -> ?⁠verbose:bool -> ?⁠clean:bool -> string -> (string * Mlpost__.Types.commandpic) list -> unit

val mp : string -> ?⁠prelude:string -> Mlpost.Defaults.jobs -> Mlpost.File.t * string Mlpost.Misc.IntMap.t

Generate files of corresponding type, using the argument of type jobs, and return information about the created files

val mps : ?⁠prelude:string -> ?⁠verbose:bool -> string -> Mlpost.Defaults.jobs -> Mlpost.File.t list

val pdf : ?⁠prelude:string -> ?⁠verbose:bool -> string -> Mlpost.Defaults.jobs -> Mlpost.File.t list

val png : ?⁠prelude:string -> ?⁠verbose:bool -> string -> Mlpost.Defaults.jobs -> Mlpost.File.t list

val temp_mp : ?⁠prelude:string -> ?⁠verbose:bool -> ?⁠clean:bool -> string -> Mlpost.Defaults.jobs -> Mlpost.File.t * string Mlpost.Misc.IntMap.t

Same as above, but use a temporary directory

val temp_mps : ?⁠prelude:string -> ?⁠verbose:bool -> ?⁠clean:bool -> string -> Mlpost.Defaults.jobs -> Mlpost.File.t list

val temp_pdf : ?⁠prelude:string -> ?⁠verbose:bool -> ?⁠clean:bool -> string -> Mlpost.Defaults.jobs -> Mlpost.File.t list

val temp_png : ?⁠prelude:string -> ?⁠verbose:bool -> ?⁠clean:bool -> string -> Mlpost.Defaults.jobs -> Mlpost.File.t list

val dump_mp : ?⁠prelude:string -> ?⁠verbose:bool -> ?⁠clean:bool -> string -> unit

Same as above, but use a temporary directory and take jobs from the job queue

val dump_mps : ?⁠prelude:string -> ?⁠verbose:bool -> ?⁠clean:bool -> string -> unit

val dump_png : ?⁠prelude:string -> ?⁠verbose:bool -> ?⁠clean:bool -> string -> unit

val dump_pdf : ?⁠prelude:string -> ?⁠verbose:bool -> ?⁠clean:bool -> string -> unit

val dump_tex : ?⁠prelude:string -> string -> unit

val slideshow : Mlpost__.Types.commandpic list -> int -> (int * Mlpost__.Types.commandpic) list

val emit_slideshow : string -> Mlpost__.Types.commandpic list -> unit

val dumpable : unit -> unit

val depend : string -> unit

val read_prelude_from_tex_file : string -> string

val emit : string -> Mlpost__.Types.commandpic -> unit

val dump : ?⁠prelude:string -> ?⁠pdf:bool -> ?⁠eps:bool -> ?⁠verbose:bool -> ?⁠clean:bool -> string -> unit

val set_filename_prefix : string -> unit

val figures_names : unit -> string list

type job = Mlpost.Defaults.job

type jobs = Mlpost.Defaults.jobs