Module Mlpost__Metapost_tool

val read_prelude_from_tex_file : string -> string

val rnd_state : Stdlib.Random.State.t

val create_temp_dir : string -> string -> Mlpost.File.Dir.t

create a temporary directory

val tempdir : ?⁠clean:bool -> string -> string -> (Mlpost.File.Dir.t -> Mlpost.File.Dir.t -> 'a * Mlpost.File.t list) -> 'a

create a temporary directory and call function f within. * f has to return a result res and a list of files files that are * copied from the tmpdir to the workdir

val append_dir : string -> string -> string