Module Mlpost_graphics.Double

Create mlpost figures and call graphics function

val clear_graph : unit -> unit

type color

val black : color

val white : color

val red : color

val green : color

val blue : color

val moveto : int -> int -> unit

Position the current point.

val lineto : int -> int -> unit

Draw a line with endpoints the current point and the given point, and move the current point to the given point.

val draw_rect : int -> int -> int -> int -> unit

draw_rect x y w h draws the rectangle with lower left corner at x,y, width w and height h. The current point is unchanged. Raise Invalid_argument if w or h is negative.

Text drawing

val draw_string : string -> unit

Draw a character or a character string with lower left corner at current position. After drawing, the current position is set to the lower right corner of the text drawn.

val set_color : color -> unit

Filling

val fill_circle : int -> int -> int -> unit

Fill a circle with the current color. The parameters are the same as for Graphics.draw_circle.

val get_cmds : unit -> Mlpost.Command.t

val set_emit : string -> unit

val synchronize : unit -> unit